Feat:sidebar 안에 search 컴포넌트합치기
