+++ src/components/SideBar/SideBar.jsx
@@ -1,10 +1,12 @@
 import { useQuery } from '@tanstack/react-query';
 import { SideBarButton, SideBarContainer, SideBarMenu, SideBarMenuItem } from './SidBarStyledcomponents';
-import supabase from './supabaseClient';
-import { useState } from 'react';
+import supabase from '../../supabase/supabaseClient';
+import { useEffect, useState } from 'react';
+import Search from './Search';
 
 const SideBar = () => {
   const [isOpen, setIsOpen] = useState(false);
+  const [filteredShops, setFilteredShops] = useState([]);
 
   const toggleSidebar = () => {
     setIsOpen(!isOpen);
@@ -22,8 +24,15 @@ const SideBar = () => {
     queryKey: ['shops'],
     queryFn: fetchRestaurants
   });
+
+  useEffect(() => {
+    if (shops) {
+      setFilteredShops(shops);
+    }
+  }, [shops]);
+
   if (isPending) {
-    <div>loding..</div>;
+    <div>loading..</div>;
     return;
   }
   if (isError) {
@@ -34,19 +43,22 @@ const SideBar = () => {
   return (
     <SideBarContainer isOpen={isOpen}>
       <SideBarButton onClick={toggleSidebar}>{isOpen ? '👈' : '👉'}</SideBarButton>
-      <label htmlFor="">검색창</label>
-      <input type="text" />
+      <Search shops={shops} setFilteredShops={setFilteredShops} />
       <SideBarMenu>
-        {shops.map((shop) => (
-          <SideBarMenuItem key={shop.id} to={`/detail/${shop.id}`}>
-            이름:{shop.name}
-            장르:{shop.genre}
-            별점:{shop.rating}
-            주소:{shop.address}
-            위치:{shop.loaction}
-            사진:{shop.img}
-          </SideBarMenuItem>
-        ))}
+        {filteredShops.length > 0 ? (
+          filteredShops.map((shop) => (
+            <SideBarMenuItem key={shop.id} to={`/detail/${shop.id}`}>
+              이름:{shop.name}
+              장르:{shop.genre}
+              별점:{shop.rating}
+              주소:{shop.address}
+              위치:{shop.loaction}
+              사진:{shop.img}
+            </SideBarMenuItem>
+          ))
+        ) : (
+          <p>검색 결과가 없습니다.</p>
+        )}
       </SideBarMenu>
     </SideBarContainer>
   );
