+++ src/pages/MainPage/MainPage.jsx
@@ -1,10 +1,9 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import Map from '../../components/Map';
-import Search from '../../components/Search';
+import Search from '../../components/SideBar/Search';
 import Detail from '../DetailPage/Detail';
 import SideBar from '../../components/SideBar/SideBar';
-
 
 const MainPage = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -15,7 +14,7 @@ const MainPage = () => {
 
   return (
     <StMain>
-      <Map />
+      {/* <Map /> */}
       <SideBar>
         <button onClick={toggleModal}>상세보기</button>
         <Search />
